+++ lib/goodcheck/commands/check.rb
@@ -6,20 +6,25 @@ module Goodcheck
       attr_reader :targets
       attr_reader :reporter
       attr_reader :stderr
+      attr_reader :force_download
+      attr_reader :home_path
 
       include ConfigLoading
+      include HomePath
 
-      def initialize(config_path:, rules:, targets:, reporter:, stderr:)
+      def initialize(config_path:, rules:, targets:, reporter:, stderr:, home_path:, force_download:)
         @config_path = config_path
         @rules = rules
         @targets = targets
         @reporter = reporter
         @stderr = stderr
+        @force_download = force_download
+        @home_path = home_path
       end
 
       def run
         reporter.analysis do
-          load_config!
+          load_config!(force_download: force_download, cache_path: cache_dir_path)
           each_check do |buffer, rule|
             reporter.rule(rule) do
               analyzer = Analyzer.new(rule: rule, buffer: buffer)
@@ -46,25 +51,32 @@ module Goodcheck
 
       def each_check
         targets.each do |target|
-          each_file target, immediate: true do |path|
-            reporter.file(path) do
-              buffers = {}
+          Goodcheck.logger.info "Checking target: #{target}"
+          Goodcheck.logger.tagged target.to_s do
+            each_file target, immediate: true do |path|
+              Goodcheck.logger.debug "Checking file: #{path}"
+              Goodcheck.logger.tagged path.to_s do
+                reporter.file(path) do
+                  buffers = {}
 
-              config.rules_for_path(path, rules_filter: rules) do |rule, glob|
-                begin
-                  encoding = glob&.encoding || Encoding.default_external.name
+                  config.rules_for_path(path, rules_filter: rules) do |rule, glob|
+                    Goodcheck.logger.debug "Checking rule: #{rule.id}"
+                    begin
+                      encoding = glob&.encoding || Encoding.default_external.name
 
-                  if buffers[encoding]
-                    buffer = buffers[encoding]
-                  else
-                    content = path.read(encoding: encoding).encode(Encoding.default_internal || Encoding::UTF_8)
-                    buffer = Buffer.new(path: path, content: content)
-                    buffers[encoding] = buffer
+                      if buffers[encoding]
+                        buffer = buffers[encoding]
+                      else
+                        content = path.read(encoding: encoding).encode(Encoding.default_internal || Encoding::UTF_8)
+                        buffer = Buffer.new(path: path, content: content)
+                        buffers[encoding] = buffer
+                      end
+
+                      yield buffer, rule
+                    rescue ArgumentError => exn
+                      stderr.puts "#{path}: #{exn.inspect}"
+                    end
                   end
-
-                  yield buffer, rule
-                rescue ArgumentError => exn
-                  stderr.puts "#{path}: #{exn.inspect}"
                 end
               end
             end